+++ lib/rubocop/cop/rspec_structure_cops.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
+require_relative "rspec_structure/asymmetric_contexts"
 require_relative "rspec_structure/condition_in_example"
 require_relative "rspec_structure/multiple_examples_in_group"